Update monitoring API
