+++ src/main/java/org/craftercms/engine/controller/rest/MonitoringController.java
@@ -1,0 +1,45 @@
+/*
+ * Copyright (C) 2007-2019 Crafter Software Corporation. All Rights Reserved.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package org.craftercms.engine.controller.rest;
+
+import java.util.HashMap;
+import java.util.List;
+
+import org.craftercms.commons.monitoring.rest.MonitoringRestControllerBase;
+import org.craftercms.engine.util.logging.CircularQueueLogAppender;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+
+/**
+ * Rest controller to provide monitoring information & site logs
+ */
+@RestController
+@RequestMapping(MonitoringController.URL_ROOT)
+public class MonitoringController extends MonitoringRestControllerBase {
+
+    public final static String URL_ROOT = "/api/1";
+    public final static String LOG_URL = "/log";
+
+    @GetMapping(LOG_URL)
+    public List<HashMap<String,Object>> getLoggedEvents(@RequestParam String site, @RequestParam long since) {
+        return CircularQueueLogAppender.loggerQueue().getLoggedEvents(site, since);
+    }
+
+}
